requirements.txt
django
psycopg2-binary
SQLAlchemy
drf-yasg
django-environ
djangorestframework-simplejwt
django-rest-framework
django-filter
